requirements.txt
django
djangorestframework
markdown
django-filter
django-crispy-forms
coreapi
django-cors-headers
django-debug-toolbar
psycopg2